refactor: extract shared loadAccreditationSources helper
Both run-stream-promotion and run-balance-divergence-diagnostic loaded
the same organisation/registration/accreditation data with identical
lookup logic. Extract the shared ~50 lines into a single helper in
load-accreditation-sources.js and have both callers use it.
